fix: rename es-* configs to es2025-* for year-based naming
Rename JavaScript-only presets to include ECMAScript year version:
- es-browser.js → es2025-browser.js
- es-node.js → es2025-node.js
- es-react.js → es2025-react.js
- es-*-disable-styles.js → es2025-*-disable-styles.js

This follows the plan's stated intention to use year-based naming
(ES2025) instead of edition numbers (ES8), consistent with the
ECMAScript standard's own naming convention.

Updated:
- Package.json exports map
- Internal imports in configuration files
