+++ configurations/es2025-browser-disable-styles.js
@@ -1,15 +1,14 @@
 import { defineConfig } from "eslint/config";
-import esNode from "./es-node.js";
+import esBrowser from "./es2025-browser.js";
 import bestPractices from "../disable-styles/best-practices.js";
 import errors from "../disable-styles/errors.js";
 import style from "../disable-styles/style.js";
 import es2025 from "../disable-styles/es2025.js";
 import imports from "../disable-styles/imports.js";
-import nodeStyles from "../disable-styles/node.js";
 import variables from "../disable-styles/variables.js";
 
 export default defineConfig([
-  ...esNode,
+  ...esBrowser,
   {
     rules: {
       ...bestPractices,
@@ -17,7 +16,6 @@ export default defineConfig([
       ...style,
       ...es2025,
       ...imports,
-      ...nodeStyles,
       ...variables,
     },
   },
