+++ configurations/es2025-browser.js
