+++ configurations/es2025-node-disable-styles.js
@@ -1,17 +1,15 @@
 import { defineConfig } from "eslint/config";
-import esReact from "./es-react.js";
+import esNode from "./es2025-node.js";
 import bestPractices from "../disable-styles/best-practices.js";
 import errors from "../disable-styles/errors.js";
 import style from "../disable-styles/style.js";
 import es2025 from "../disable-styles/es2025.js";
 import imports from "../disable-styles/imports.js";
+import nodeStyles from "../disable-styles/node.js";
 import variables from "../disable-styles/variables.js";
-import reactStyles from "../disable-styles/react.js";
-import reactA11yStyles from "../disable-styles/react-a11y.js";
-import reactHooksStyles from "../disable-styles/react-hooks.js";
 
 export default defineConfig([
-  ...esReact,
+  ...esNode,
   {
     rules: {
       ...bestPractices,
@@ -19,10 +17,8 @@ export default defineConfig([
       ...style,
       ...es2025,
       ...imports,
+      ...nodeStyles,
       ...variables,
-      ...reactStyles,
-      ...reactA11yStyles,
-      ...reactHooksStyles,
     },
   },
 ]);
