+++ configurations/es2025-react-disable-styles.js
@@ -1,14 +1,17 @@
 import { defineConfig } from "eslint/config";
-import esBrowser from "./es-browser.js";
+import esReact from "./es2025-react.js";
 import bestPractices from "../disable-styles/best-practices.js";
 import errors from "../disable-styles/errors.js";
 import style from "../disable-styles/style.js";
 import es2025 from "../disable-styles/es2025.js";
 import imports from "../disable-styles/imports.js";
 import variables from "../disable-styles/variables.js";
+import reactStyles from "../disable-styles/react.js";
+import reactA11yStyles from "../disable-styles/react-a11y.js";
+import reactHooksStyles from "../disable-styles/react-hooks.js";
 
 export default defineConfig([
-  ...esBrowser,
+  ...esReact,
   {
     rules: {
       ...bestPractices,
@@ -17,6 +20,9 @@ export default defineConfig([
       ...es2025,
       ...imports,
       ...variables,
+      ...reactStyles,
+      ...reactA11yStyles,
+      ...reactHooksStyles,
     },
   },
 ]);
